.gemini/skills/skill-comply/pyproject.toml
[project]
name = "skill-comply"
version = "0.1.0"
description = "Automated skill compliance measurement for Claude Code"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
